+++ Projects/App/Sources/Journeys/ClaimsJourney.swift
@@ -39,7 +39,11 @@ extension AppJourney {
         if hAnalyticsExperiment.claimsFlow {
             if hAnalyticsExperiment.claimsTriaging {
                 ClaimJourneys.showClaimEntrypointsNew(origin: origin) { newOrigin in
-                    honestyPledge(from: newOrigin)
+                    ClaimJourneys.showClaimEntrypointType(origin: newOrigin) { origin in
+                        ClaimJourneys.showClaimEntrypointOption(origin: origin) { newOrigin in
+                            honestyPledge(from: newOrigin)
+                        }
+                    }
                 }
             } else {
                 ClaimJourneys.showClaimEntrypointsOld(origin: origin) { newOrigin in
