+++ Projects/Claims/Sources/Journeys/ClaimJourneys.swift
@@ -289,9 +289,71 @@ public class ClaimJourneys {
     ) -> some JourneyPresentation {
         HostingJourney(
             SubmitClaimStore.self,
-            rootView: SelectClaimEntrypoint(),
+            rootView: SelectClaimEntrypointGroup(
+                selectedEntrypoints: { entrypoints in
+                    let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                    store.send(.setSelectedEntrypoints(entrypoints: entrypoints))
+                }),
             //            style: .detented(.large, modally: true)
             style: .modal
+        ) { action in
+            if case let .commonClaimOriginSelected(origin) = action {
+                GroupJourney { context in
+                    switch origin {
+                    case .generic:
+                        ContinueJourney()
+                    case let .commonClaims(id):
+                        redirectJourney(ClaimsOrigin.commonClaims(id: id))
+                    case let .commonClaimsWithOption(id, optionId):
+                        redirectJourney(ClaimsOrigin.commonClaimsWithOption(id: id, optionId: optionId))
+                    }
+                }
+            }
+        }
+    }
+
+    @JourneyBuilder
+    public static func showClaimEntrypointType(
+        origin: ClaimsOrigin,
+        @JourneyBuilder redirectJourney: @escaping (_ newOrigin: ClaimsOrigin) -> some JourneyPresentation
+    ) -> some JourneyPresentation {
+
+        HostingJourney(
+            SubmitClaimStore.self,
+            rootView: SelectClaimEntrypointType(selectedEntrypointOptions: { options, selectedEntrypoint in
+                let store: SubmitClaimStore = globalPresentableStoreContainer.get()
+                store.send(.setSelectedEntrypointOptions(entrypoints: options))
+                store.send(.setSelectedEntrypointId(entrypoints: selectedEntrypoint))
+            })
+            //            style: .detented(.large, modally: true)
+            //            style: .modal
+        ) { action in
+            if case let .commonClaimOriginSelected(origin) = action {
+                GroupJourney { context in
+                    switch origin {
+                    case .generic:
+                        ContinueJourney()
+                    case let .commonClaims(id):
+                        redirectJourney(ClaimsOrigin.commonClaims(id: id))
+                    case let .commonClaimsWithOption(id, optionId):
+                        redirectJourney(ClaimsOrigin.commonClaimsWithOption(id: id, optionId: optionId))
+                    }
+                }
+            }
+        }
+    }
+
+    @JourneyBuilder
+    public static func showClaimEntrypointOption(
+        origin: ClaimsOrigin,
+        @JourneyBuilder redirectJourney: @escaping (_ newOrigin: ClaimsOrigin) -> some JourneyPresentation
+    ) -> some JourneyPresentation {
+
+        HostingJourney(
+            SubmitClaimStore.self,
+            rootView: SelectClaimEntrypointOption()
+                //            style: .detented(.large, modally: true)
+                //            style: .modal
         ) { action in
             if case let .commonClaimOriginSelected(origin) = action {
                 GroupJourney { context in
